feat: update and delete apis

+++ src/apis/favorite/schemas.go
@@ -5,6 +5,6 @@ type AddFavoriteModel struct {
 }
 
 type PriceLimitModel struct {
-	PriceLimit int `json:"price_limit"`
-	ItemId     int `json:"item_id"`
+	PriceLimit float32 `json:"price_limit"`
+	ItemId     int     `json:"item_id"`
 }
