+++ src/apis/item/schemas.go
@@ -3,7 +3,7 @@ package item
 type CreateItemModel struct {
 	CommodityID int     `json:"commodity_id" validate:"required"`
 	PlatformID  int     `json:"platform_id" validate:"required"`
-	Price       float64 `json:"price" validate:"required"`
+	Price       float32 `json:"price" validate:"required"`
 	ItemName    string  `json:"item_name" validate:"required"`
 }
 
@@ -13,5 +13,5 @@ type CreateItemModel struct {
 type UpdateItemModel struct {
 	CommodityItemID int     `json:"commodity_item_id" validate:"required"`
 	ItemName        string  `json:"item_name" validate:"required"`
-	Price           float64 `json:"price" validate:"required"`
+	Price           float32 `json:"price" validate:"required"`
 }
